+++ knytt/juni/ClimbCheckers.cs
@@ -2,15 +2,15 @@ using Godot;
 
 public class ClimbCheckers : Node2D
 {
-    Area2D right_checker;
-    Area2D left_checker;
+    public Area2D RightChecker { get; private set; }
+    public Area2D LeftChecker { get; private set; }
 
-    public bool RightColliding { get { return right_checker.GetOverlappingBodies().Count > 0 && right_checker.GetOverlappingAreas().Count == 0; } }
-    public bool LeftColliding { get { return left_checker.GetOverlappingBodies().Count > 0 && left_checker.GetOverlappingAreas().Count == 0; } }
+    public bool RightColliding { get { return RightChecker.GetOverlappingBodies().Count > 0 && RightChecker.GetOverlappingAreas().Count == 0; } }
+    public bool LeftColliding { get { return LeftChecker.GetOverlappingBodies().Count > 0 && LeftChecker.GetOverlappingAreas().Count == 0; } }
 
     public override void _Ready()
     {
-        right_checker = GetNode<Area2D>("RightChecker");
-        left_checker = GetNode<Area2D>("LeftChecker");
+        RightChecker = GetNode<Area2D>("RightChecker");
+        LeftChecker = GetNode<Area2D>("LeftChecker");
     }
 }