+++ knytt/objects/GDKnyttObjectFactory.cs
@@ -62,6 +62,7 @@ public static class GDKnyttObjectFactory
         ObjectLookup[new KnyttPoint(4, 18)] = "ToastMonster";
         ObjectLookup[new KnyttPoint(7, 8)] =  "Rain";
         ObjectLookup[new KnyttPoint(7, 9)] =  "RaindropObject";
+        ObjectLookup[new KnyttPoint(7, 16)] =  "Cloud";
         ObjectLookup[new KnyttPoint(10, 5)] =  "BouncerEnemy";
         ObjectLookup[new KnyttPoint(12, 5)] =  "GhostBlock";
         ObjectLookup[new KnyttPoint(12, 11)] =  "Ghost";