+++ knytt/objects/banks/bank7/Cloud.cs
@@ -1,0 +1,10 @@
+using Godot;
+
+public class Cloud : GDKnyttBaseObject, IParticleFetch
+{
+    public int ParticleType { get { return 0; } }
+
+    protected override void _impl_initialize() { }
+
+    protected override void _impl_process(float delta) { }
+}