+++ Changelog.md
@@ -13,7 +13,9 @@ All notable changes to YKnytt will be documented here.
 - Spring
 - Updraft
 - Cat
-- "Sky Flowerz" and "Gustav's Daughter" worlds now built-in
+- "Sky Flower" and "Gustav's Daughter" worlds now built-in
+- Walk / Climb particle spawning
+- Cloud object
 
 ## Fixed
 - Rain drops now spawn down a bit (They were colliding with the tile above)